Switched to pop and swap method of entity removal
Removed openIdxs slice in archetype, and now relying on a copy of the
entity at the end of the compstore to be able to pop and swap.

Switched to pop and swap method of entity removal

Removed openIdxs slice in archetype, and now relying on a copy of the
entity at the end of the compstore to be able to pop and swap.